refactor: pure domain services (money/dates/split) in TypeScript
