+++ src/services/dates.js
@@ -1,13 +1,2 @@
-function monthOf(date) {
-  return String(date).slice(0, 7);
-}
-
-function addMonths(ym, n) {
-  const [y, m] = ym.split('-').map(Number);
-  const total = y * 12 + (m - 1) + n;
-  const ny = Math.floor(total / 12);
-  const nm = (total % 12) + 1;
-  return `${ny}-${String(nm).padStart(2, '0')}`;
-}
-
-module.exports = { monthOf, addMonths };
+// Migration shim: pure logic now lives in src/domain/services/dates.ts.
+module.exports = require('../domain/services/dates');
